refactor: reorganize package manager utilities and update import paths
